Fix issue #43 (#49)
* add travis file

* Wrangle Travis

 - add in a script for building Quick DER including dependencies
 - let Travis run the script, instead of putting the whole build
   instruction in the .travis.yml file.

* Wrangle Travis

 - Be more verbose in building, and quicker to error out.

* enable root

* sudo the install

* install the python module

* install in install step

* echo commands

* weird shell

* use https so git doesnt ask key question

* move setup.py to root

* rename to pyquickder

* build from root

* make working with asn1ate-0.5.1.dev0

* add a new test

* fix remaining issues

* disable test for now (issue #42)

+++ python/quick_der/main.py
@@ -193,6 +193,7 @@ class QuickDER2c(QuickDERgeneric):
 
         self.semamod = semamod
         self.refmods = refmods
+
         # Open the output file
         super(QuickDER2c, self).__init__(outfn, '.h')
         # Setup function maps
@@ -261,24 +262,31 @@ class QuickDER2c(QuickDERgeneric):
         self.writeln()
         closer = ''
         rmfns = set()
-        for rm in self.semamod.imports.symbols_imported.keys():
-            rmfns.add(tosym(rm.rsplit('.', 1)[0]).lower())
+
+        if not self.semamod.imports:
+            return
+
+        for rm in self.semamod.imports.imports.keys():
+            rmfns.add(tosym(str(rm).rsplit('.', 1)[0]).lower())
+
         for rmfn in rmfns:
             self.writeln('#include <quick-der/' + rmfn + '.h>')
             closer = '\n\n'
         self.write(closer)
         closer = ''
-        for rm in self.semamod.imports.symbols_imported.keys():
-            rmfn = tosym(rm.rsplit('.', 1)[0]).lower()
-            for sym in self.semamod.imports.symbols_imported[rm]:
+
+        for rm in self.semamod.imports.imports.keys():
+            rmfn = tosym(str(rm).rsplit('.', 1)[0]).lower()
+            for sym in self.semamod.imports.imports[rm]:
                 self.writeln('typedef DER_OVLY_' + tosym(rmfn) + '_' + tosym(sym) + ' DER_OVLY_' + tosym(
                     self.unit) + '_' + tosym(sym) + ';')
                 closer = '\n\n'
         self.write(closer)
         closer = ''
-        for rm in self.semamod.imports.symbols_imported.keys():
-            rmfn = tosym(rm.rsplit('.', 1)[0]).lower()
-            for sym in self.semamod.imports.symbols_imported[rm]:
+
+        for rm in self.semamod.imports.imports.keys():
+            rmfn = tosym(str(rm).rsplit('.', 1)[0]).lower()
+            for sym in self.semamod.imports.imports[rm]:
                 self.writeln(
                     '#define DER_PIMP_' + tosym(self.unit) + '_' + tosym(sym) + '(implicit_tag) DER_PIMP_' + tosym(
                         rmfn) + '_' + tosym(sym) + '(implicit_tag)')
@@ -338,7 +346,7 @@ class QuickDER2c(QuickDERgeneric):
             fld = ''
         else:
             fld = '_' + fld
-        # OLD:TEST:TODO# mod = node.module_name or self.unit
+        # OLD:TEST:TODO# mod = node.module_ref or self.unit
         mod = self.unit
         self.comma()
         self.writeln('const struct der_subparser_action DER_PSUB_' + mod + '_' + tp + fld + ' [] = { \\')
@@ -431,12 +439,12 @@ class QuickDER2c(QuickDERgeneric):
         return []
 
     def overlayDefinedType(self, node, tp, fld):
-        mod = node.module_name or self.unit
+        mod = node.module_ref or self.unit
         self.write('DER_OVLY_' + tosym(mod) + '_' + tosym(node.type_name))
 
     def packDefinedType(self, node, implicit=False, outer_tag=None):
         # There should never be anything of interest in outer_tag
-        mod = node.module_name or self.unit
+        mod = node.module_ref or self.unit
         self.comma()
         if outer_tag is None:
             tagging = 'DER_PACK_'
@@ -448,18 +456,18 @@ class QuickDER2c(QuickDERgeneric):
 
     def psubDefinedType(self, node, tp, fld, prim):
         dprint('DefinedType type:', node.type_name, '::', type(node.type_name))
-        modnm = node.module_name
-        dprint('AFTER modnm = node.module_name', modnm)
-        if modnm is None:
-            syms = self.semamod.imports.symbols_imported
+        modnm = node.module_ref
+        dprint('AFTER modnm = node.module_ref', modnm)
+        if not modnm and self.semamod.imports:
+            syms = self.semamod.imports.imports
             dprint('SYMS.KEYS() =', syms.keys())
             for mod in syms.keys():
                 if node.type_name in syms[mod]:
-                    modnm = mod.lower()
-                    dprint('AFTER modnm = mod.lower ()', modnm)
+                    modnm = str(mod).lower()
+                    dprint('AFTER modnm = str(mod).lower ()', modnm)
                     break
         if modnm is None:
-            # NOT_IN_GENERAL# modnm = node.module_name
+            # NOT_IN_GENERAL# modnm = node.module_ref
             modnm = self.unit.lower()
             dprint('AFTER modnm = self.unit.lower ()', modnm)
             dprint('MODNM =', modnm, '::', type(modnm))
@@ -515,7 +523,7 @@ class QuickDER2c(QuickDERgeneric):
             self.comma()
             self.write('DER_PACK_ENTER | ' + outer_tag)
         mytag = 'DER_TAG_' + (node.class_name or 'CONTEXT') + '(' + node.class_number + ')'
-        if self.semamod.resolve_tag_implicity(node.implicity, node.type_decl) == TagImplicitness.IMPLICIT:
+        if self.semamod.resolve_tag_implicitness(node.implicitness, node.type_decl) == TagImplicitness.IMPLICIT:
             self.generate_pack_node(node.type_decl, implicit=False, outer_tag=mytag)
         else:
             self.comma()
@@ -531,7 +539,7 @@ class QuickDER2c(QuickDERgeneric):
         if not implicit:
             self.comma()
             self.write('DER_PACK_ENTER | DER_TAG_' + (node.class_name or 'CONTEXT') + '(' + node.class_number + ')')
-        implicit_sub = (self.semamod.resolve_tag_implicity(node.implicity, node.type_decl) == TagImplicitness.IMPLICIT)
+        implicit_sub = (self.semamod.resolve_tag_implicitness(node.implicitness, node.type_decl) == TagImplicitness.IMPLICIT)
         self.generate_pack_node(node.type_decl, implicit=implicit_sub)
         if not implicit:
             self.comma()
@@ -822,9 +830,13 @@ class QuickDER2py(QuickDERgeneric):
         self.writeln()
         self.writeln('import quick_der.api as ' + api_prefix)
         self.writeln()
-        imports = self.semamod.imports.symbols_imported
+
+        if not self.semamod.imports:
+            return
+
+        imports = self.semamod.imports.imports
         for rm in imports.keys():
-            pymod = tosym(rm.rsplit('.', 1)[0]).lower()
+            pymod = tosym(str(rm).rsplit('.', 1)[0]).lower()
             self.write('from ' + pymod + ' import ')
             self.writeln(', '.join(map(tosym, imports[rm])))
         self.writeln()
@@ -1008,12 +1020,12 @@ class QuickDER2py(QuickDERgeneric):
         return self.funmap_pytype[tnm](node, **subarg)
 
     def pytypeDefinedType(self, node, **subarg):
-        modnm = node.module_name
-        if modnm is None:
-            syms = self.semamod.imports.symbols_imported
+        modnm = node.module_ref
+        if not modnm and self.semamod.imports:
+            syms = self.semamod.imports.imports
             for mod in syms.keys():
                 if node.type_name in syms[mod]:
-                    modnm = mod.lower()
+                    modnm = str(mod).lower()
                     break
         if modnm is None:
             modnm = self.unit.lower()
@@ -1062,7 +1074,7 @@ class QuickDER2py(QuickDERgeneric):
 
     def pytypeTagged(self, node, implicit_tag=None):
         mytag = 'DER_TAG_' + (node.class_name or 'CONTEXT') + '(' + node.class_number + ')'
-        if self.semamod.resolve_tag_implicity(node.implicity, node.type_decl) == TagImplicitness.IMPLICIT:
+        if self.semamod.resolve_tag_implicitness(node.implicitness, node.type_decl) == TagImplicitness.IMPLICIT:
             # Tag implicitly by handing mytag down to type_decl
             (pck, recp) = self.generate_pytype(node.type_decl,
                                                implicit_tag=mytag)
@@ -1247,12 +1259,12 @@ class QuickDER2testdata(QuickDERgeneric):
         return self.funmap_tdgen[tnm](node, **subarg)
 
     def tdgenDefinedType(self, node, **subarg):
-        modnm = node.module_name
+        modnm = node.module_ref
         if modnm is None:
-            syms = self.semamod.imports.symbols_imported
+            syms = self.semamod.imports.imports
             for mod in syms.keys():
                 if node.type_name in syms[mod]:
-                    modnm = mod.lower()
+                    modnm = str(mod).lower()
                     break
         if modnm is None:
             modnm = self.unit.lower()
@@ -1329,7 +1341,7 @@ class QuickDER2testdata(QuickDERgeneric):
     def tdgenTagged(self, node, implicit_tag=None):
         # Tagged values delegate to type_decl, prefixing a header
         (subcnt, subgen) = self.generate_tdgen(node.type_decl)
-        am_implicit = self.semamod.resolve_tag_implicity(node.implicity, node.type_decl) == TagImplicitness.IMPLICIT
+        am_implicit = self.semamod.resolve_tag_implicitness(node.implicitness, node.type_decl) == TagImplicitness.IMPLICIT
         tag = self.nodeclass2basaltag[node.class_name or 'CONTEXT']
         tag |= int(node.class_number)
 
@@ -1588,8 +1600,12 @@ def main(script_name, script_args):
     imports = list(refmods.keys())
     while len(imports) > 0:
         dm = refmods[imports.pop(0).lower()]
-        for rm in dm.imports.symbols_imported.keys():
-            rm = rm.lower()
+
+        if not dm.imports:
+            continue
+
+        for rm in dm.imports.imports.keys():
+            rm = str(rm).lower()
             if rm not in refmods:
                 dprint('Importing ASN.1 include for "%s"', rm)
                 modfh = None
@@ -1600,7 +1616,7 @@ def main(script_name, script_args):
                     except IOError:
                         continue
                 if modfh is None:
-                    raise Exception('No include file "%s.asn1" found', rm)
+                    raise Exception('No include file "{}.asn1" found'.format(rm))
                 asn1txt = modfh.read()
                 asn1tree = parser.parse_asn1(asn1txt)
                 dprint('Building semantic model for "%s"', rm)
